Final swoop

+++ Project2/Fibonacci/Bottom_Up/Bottom_Up_Fibonacci.h
@@ -8,9 +8,9 @@ class BottomUpFib : public Fibonacci
 public:
 	BottomUpFib() : Fibonacci("Bottom Up Fibonacci") {};
 
-	int BUF(int n);
+	unsigned long long BUF(int n);
 
 	void RunAlgorithm();
 
-	void FileWriter(int iteration, size_t input[], size_t output[], std::chrono::time_point<std::chrono::steady_clock> start[], std::chrono::time_point<std::chrono::steady_clock> stop[]);
+	void FileWriter(int iteration, size_t input[], unsigned long long output[], std::chrono::time_point<std::chrono::steady_clock> start[], std::chrono::time_point<std::chrono::steady_clock> stop[]);
 };
